class bouteille Biere crée

+++ TP2_Bieres_Lacombe_Barberan/src/tp2_bieres_lacombe_barberan/BouteilleBiere.java
@@ -10,7 +10,7 @@ package tp2_bieres_lacombe_barberan;
  */
 public class BouteilleBiere {
     String nom; 
-    float  degreAlcool;
+    double  degreAlcool;
     String  brasserie;
     boolean ouverte;
